% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_level.R
\name{adjustment_step_parallel}
\alias{adjustment_step_parallel}
\title{Adjust a hierarchy level in parallel}
\usage{
adjustment_step_parallel(data, mod, combatmode, method)
}
\arguments{
\item{data}{Matrix or dataframe in the format (samples, features).
Additional column names are \"Batch\", \"Cov_X\" (were X may be any number),
\"Label\" and \"Sample\".}

\item{mod}{Dataframe with potential covariables to use. May be emty.}

\item{combatmode}{Integer, encoding the parameters to use for ComBat.
1 (default)    par.prior = TRUE, mean.only = FALSE
2              par.prior = TRUE, mean.only = TRUE
3              par.prior = FALSE, mean.only = FALSE
4              par.prior = FALSE, mean.only = TRUE
Will be ignored, if method=="limma".}

\item{method}{Adjustment method to use. Should either be \"ComBat\" or \"limma\".}
}
\value{
A matrix/dataframe mirroring the shape of the input. The data will
be batch-effect adjusted by BERT.
}
\description{
This function uses ComBat or limma to adjust an entire hierarchy level.
}
