% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{bootstrap.cpue_area}
\alias{bootstrap.cpue_area}
\title{Bootstrap CPUE by seabed area}
\usage{
\method{bootstrap}{cpue_area}(x, nboot = 10000, ...)
}
\arguments{
\item{x}{object of class cpue_area}

\item{nboot}{number of bootstrap samples (default=10000)}

\item{...}{additional arguments}
}
\description{
Bootstrap CPUE by seabed area
}
