% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bootstrap}
\alias{bootstrap}
\title{S3 generic bootstrap method}
\usage{
bootstrap(x, nboot, ...)
}
\arguments{
\item{x}{an object of of the appropriate class}

\item{nboot}{number of bootstrap samples}

\item{...}{additional parameters}
}
\description{
Estimate confidence intervals using a non-parametric bootstrap.
}
