% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_data_sim_RB}
\alias{catch_data_sim_RB}
\title{Simulated catch data from hypothetical Research blocks for vignette examples}
\format{A dataframe with 400 rows and 8 columns
\describe{
\item{ID}{identifies a unique fishing event - i.e. both longline set and haul information}
\item{Season}{The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) that the longline was set and hauled}
\item{CRUISE_ID}{An ID used by Observers that is recorded with tagged fish releases}
\item{SET_ID}{An ID used by Observers that is recorded for each set-haul}
\item{SPECIES_CODE}{The three letter code used by CCAMLR to record a species. In the simulated data example this represent either Antarctic toothfish = "TOA" or Patagonian toothfish = "TOP"}
\item{CAUGHT_KG_TOTAL}{A simulated normally distributed catch that represents retained catch per longline haul}
\item{LINE_LENGTH}{A simulated normally distributed longline length}
\item{RESEARCH_BLOCK_CODE}{A simulated hypothetical Research block code that the data fall within}
}}
\usage{
catch_data_sim_RB
}
\description{
A dataset that aims to simulate the catch data necessary for toothfish local biomass estimates that is stored in CCAMLR's cdb database
}
\keyword{datasets}
