% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{chapman_wt}
\alias{chapman_wt}
\title{Chapman estimate of biomass}
\usage{
chapman_wt(avail_tags, recaps, catch, mean_wt = 0)
}
\arguments{
\item{avail_tags}{the available tags}

\item{recaps}{tag recaptures}

\item{catch}{the catch in the same units as mean weight}

\item{mean_wt}{mean weight of a fish in the same units as catch (default = 0)}
}
\description{
Chapman estimate of biomass
}
