% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert.R
\name{chunk_data}
\alias{chunk_data}
\title{Chunks data into n segments with (close-to) equivalent number of batches
and stores them in temporary RDS files}
\usage{
chunk_data(data, n, backend = "default")
}
\arguments{
\item{data}{Dataframe with the data to adjust}

\item{n}{The number of chunks to create}

\item{backend}{The backend to choose for communicating the data, Valid choices
are "default" and "file". The latter will use temp files for communicating
data chunks between the processes.}
}
\value{
Vector with the absolute paths to the temporary files, where the data
is stored
}
\description{
Chunks data into n segments with (close-to) equivalent number of batches
and stores them in temporary RDS files
}
