% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Dataset.R
\name{count_existing}
\alias{count_existing}
\title{Count the number of numeric features in this dataset. Columns labelled
"Batch", "Sample" or "Label" will be ignored.}
\usage{
count_existing(dataset)
}
\arguments{
\item{dataset}{Dataframe in the shape (samples, features) with optional
columns "Batch", "Sample" or "Label".}
}
\value{
Integer indicating the number of numeric values
}
\description{
Count the number of numeric features in this dataset. Columns labelled
"Batch", "Sample" or "Label" will be ignored.
}
\examples{
# generate dataset wiith 1000 features, 5 batches, 10 samples per batch and
# two genotypes
data = generateDataset(1000,5,10, 0.1, 2)
count_existing(data)
}
