% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{cpue_bio}
\alias{cpue_bio}
\title{CPUE Seabed area analogy}
\usage{
cpue_bio(fish_CPUE, fish_area, ref_CPUE, ref_area, ref_bio)
}
\arguments{
\item{fish_CPUE}{CPUE data from the research area}

\item{fish_area}{habitat area (km2) in research area}

\item{ref_CPUE}{CPUE data from reference area}

\item{ref_area}{habitat area in reference area}

\item{ref_bio}{biomass estimate in reference area}
}
\description{
Calculate biomass using the CPUE seabed area analogy
}
