% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R, R/generics.R
\name{cpue_bootstrap.cpue_area}
\alias{cpue_bootstrap.cpue_area}
\alias{cpue_bootstrap}
\alias{cpue_bootstrap.cpue_area}
\title{Estimate confidence intervals using a non-parametric bootstrap.}
\usage{
\method{cpue_bootstrap}{cpue_area}(x, nboot)

cpue_bootstrap(x, nboot)
}
\arguments{
\item{x}{an object of of the appropriate class}

\item{nboot}{number of bootstrap samples}
}
\description{
Estimate confidence intervals using a non-parametric bootstrap.
}
