% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue.R
\name{CPUE_seabed}
\alias{CPUE_seabed}
\title{CPUE Seabed area analogy}
\usage{
CPUE_seabed(fish_CPUE_data, fish_area, ref_CPUE_data, ref_area, ref_bio,
  ref_bio_cv, method = "median")
}
\arguments{
\item{fish_CPUE_data}{vector of CPUE data from the research area}

\item{fish_area}{habitat area (km2) in research area}

\item{ref_CPUE_data}{vector of CPUE data from reference area}

\item{ref_area}{habitat area (km2) in reference area}

\item{ref_bio}{biomass estimate in reference area (I assume the units are in kg)}

\item{ref_bio_cv}{CV of biomass estimate in reference area}

\item{method}{method to calculate CPUE data either "mean" or "median" (default)}
}
\description{
Calculate biomass using the CPUE seabed area analogy
}
