% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fish_weight.R
\name{est_fish_weight}
\alias{est_fish_weight}
\title{Estimate tagged fish weight from fish lengths}
\usage{
est_fish_weight(length_weight_data, length_data)
}
\arguments{
\item{length_weight_data}{is a four column dataframe with the CCAMLR ASD code,species code, length (cm) and weight (kg)}

\item{length_data}{is a six column dataframe with the CCAMLR ASD code,fishing season,cruise id, set id, species code, length (cm)}
}
\description{
Tagged fish releases and recaptures are not weighed but thier lengths are measured. 
This function uses length and weight measurements from commercial data in a log-linear model to estimate the weights of tagged fish from length data.
A dataframe is returned with all the information input into length_data plus the estimated weights
}
