% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_catch_data_cpue_est}
\alias{extract_catch_data_cpue_est}
\title{Extract haul data}
\usage{
extract_catch_data_cpue_est(catch_data, release_data, catch_seasons = NULL,
  measure, mean_fish_weight = NULL, target_species)
}
\arguments{
\item{catch_data}{catch data at haul-by-haul resolution that includes a data frame with columns of data in the following order: Unique Catch ID record reference, Fishing Season,cruise id, set id, species code, retained catch quantity (kg),long line length}

\item{release_data}{release data at the haul-by-haul resolution that includes a data frame with columns of data in the following order: Fishing Season,cruise id, set id, species code and fish weights if available and measure="weights", if fish weights are not available and measure="counts" these weights are estimated}

\item{catch_seasons}{vector of seasons of interest where the default = NULL uses all the catch data seasons included in the catch_data dataframe}

\item{measure}{"counts" if only fish counts are available or "weights" if fish weights are required}

\item{mean_fish_weight}{is the mean weight of a fish to estimate the fish release weights per haul if measure= "counts" otherwise the default=NULL}

\item{target_species}{is the species code for the species that is targetted in the research block or reference area}
}
\description{
Extract catch data for cpue by seabed area biomass
}
