% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_recaptures}
\alias{extract_recaptures}
\title{Extract recaptures}
\usage{
extract_recaptures(data, rel_seasons, recap_seasons, ...)
}
\arguments{
\item{data}{tag release and recapture data with fields SEASON_RELEASE,
SEASON_RECAPTURE and Recapture_CCAMLR_Month}

\item{rel_seasons}{vector of tag release seasons}

\item{recap_seasons}{vector of tag recapture seasons}

\item{...}{additional arguments}
}
\description{
Extract recapture data for calculation of biomass
}
