% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_recaptures_season}
\alias{extract_recaptures_season}
\title{Extract recaptures Season}
\usage{
extract_recaptures_season(data, rel_seasons)
}
\arguments{
\item{data}{tag release and recapture data with fields SEASON_RELEASE,
SEASON_RECAPTURE}

\item{rel_seasons}{vector of tag release seasons}

\item{...}{additional arguments}
}
\description{
Extract recapture data for calculation of biomass for a CCAMLR Season
}
