% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_releases}
\alias{extract_releases}
\title{Extract releases}
\usage{
extract_releases(data, seasons, ...)
}
\arguments{
\item{data}{recapture data with fields SEASON and CCAMLR_Month}

\item{seasons}{vector of seasons}

\item{...}{additional arguments}
}
\description{
Extract release data for calculation of biomass
}
