% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_Data.R
\name{format_DF}
\alias{format_DF}
\title{Format the data as expected by BERT.}
\usage{
format_DF(data)
}
\arguments{
\item{data}{Matrix or dataframe in the format (samples, features).
Additional column names are "Batch", "Cov_X" (were X may be any number),
"Label" and "Sample".}
}
\value{
The formatted matrix.
}
\description{
This function is called automatically by BERT. It removes empty columns
and removes a (usually very small) number of numeric values, if features are
unadjustable for lack of data.
}
