% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_features.R
\name{get_adjustable_features}
\alias{get_adjustable_features}
\title{Check, which features contain enough numeric data to be adjusted (at least
2 numeric values)}
\usage{
get_adjustable_features(data_batch)
}
\arguments{
\item{data_batch}{Matrix or dataframe in the format (samples, features).
Additional column names are "Batch", "Cov_X" (were X may be any number),
"Label" and "Sample".}
}
\value{
A logical with TRUE for adjustable features and FALSE for features
with too many missing values.
}
\description{
This function will be called automatically be BERT n data from each batch
independently.
}
