% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ref.R
\name{identify_adjustableFeatures_refs}
\alias{identify_adjustableFeatures_refs}
\title{Identifies the adjustable features using only the references. Similar to
the function in adjust_features.R but with different arguments}
\usage{
identify_adjustableFeatures_refs(x, batch, idx)
}
\arguments{
\item{x}{the data matrix}

\item{batch}{the list with the batches}

\item{idx}{the vector indicating whether the respective sample is to be used
as references}
}
\value{
vector indicating whether each feature can be adjusted
}
\description{
Identifies the adjustable features using only the references. Similar to
the function in adjust_features.R but with different arguments
}
