% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_weight_data_sim_RefArea}
\alias{length_weight_data_sim_RefArea}
\title{Simulated tagged fish length and weight data from hypothetical Reference Areas for vignette examples}
\format{A dataframe with 800 rows and  4 columns
\describe{
\item{ASD_CODE}{A simulated hypothetical CCAMLR Management area that represents an Area, Subarea or Division code that the data fall within}
\item{SPECIES_CODE}{The three letter code used by CCAMLR to record a species. In the simulated data example this represent either Antarctic toothfish = "TOA" or Patagonian toothfish = "TOP"}
\item{LENGTH_CM}{A simulated retained catch toothfish length}
\item{WEIGHT_KG}{A simulated retained catch weight}
}}
\usage{
length_weight_data_sim_RefArea
}
\description{
A dataset that aims to simulate the length and weight data necessary for toothfish local biomass estimates that is stored in CCAMLR's cdb database
}
\keyword{datasets}
