% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load CCAMLR data extracts}
\usage{
load_data(data_dir)
}
\arguments{
\item{data_dir}{is the directory into which the CCAMLR data extracts were
unzipped/copied to by the user}
}
\description{
Loads catch, tagged fish release, length-weight and tagged fish recapture data that have been extracted using the CCAMLRDBQueries package
}
