% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_tagr}
\alias{load_data_tagr}
\title{Load CCAMLR data extracts}
\usage{
load_data_tagr(data_dir)
}
\arguments{
\item{data_dir}{is the directory into which the CCAMLR data extracts were
unzipped/copied to by the user}
}
\description{
Load CCAMLR data extracts
}
