% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_asw.R
\name{ordinal_encode}
\alias{ordinal_encode}
\title{Ordinal encoding of a vector.}
\usage{
ordinal_encode(column)
}
\arguments{
\item{column}{The categorical vector}
}
\value{
The encoded vector
}
\description{
This function is usually called by BERT during formatting of the input.
The idea is, that Label, Batch and Covariables should only be integers
}
