% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert.R
\name{parallel_bert}
\alias{parallel_bert}
\title{Adjusts all chunks of data (in parallel) as far as possible.}
\usage{
parallel_bert(chunks, method = "ComBat", combatmode = 1, backend = "default")
}
\arguments{
\item{chunks}{vector with the filenames to the temp files where the
sub-matrices are stored}

\item{method}{the BE-correction method to use. Possible choices are ComBat
and limma}

\item{combatmode}{The mode to use for combat (ignored if limma).
Encoded options 'are the same as for HarmonizR}

\item{backend}{The backend to choose for communicating the data, Valid choices
are "default" and "file". The latter will use temp files for communicating
data chunks between the processes.}
}
\value{
dataframe with the adjusted matrix
}
\description{
Adjusts all chunks of data (in parallel) as far as possible.
}
