% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_catch}
\alias{process_catch}
\alias{process_releases}
\alias{process_recaptures}
\alias{process_releases}
\alias{process_recaptures}
\title{Process CCAMLR csv data extract}
\usage{
process_catch(data, location, species, seasons, select = NULL, ...)

process_releases(data, location, species, seasons, select = NULL, ...)

process_recaptures(data, location, species, select = NULL, ...)
}
\arguments{
\item{data}{CCAMLR data extract}

\item{location}{either a ASD code or a research block (must be character format)}

\item{species}{either "TOA" or "TOP"}

\item{seasons}{vector of CCAMLR fishing seasons}

\item{select}{fields to select in the data (default = NULL selects all fields)}

\item{...}{additional arguments}
}
\description{
Process CCAMLR csv data extract for use in Chapman estimate of biomass
}
