% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_date}
\alias{process_date}
\title{Converts a field to a date in R}
\usage{
process_date(data, column, prefix = NULL, ...)
}
\arguments{
\item{data}{CCAMLR data extract}

\item{column}{name of the field to convert to a date}

\item{prefix}{a prefix to be added to the new field names}

\item{...}{additional arguments}

\item{splitdate}{if TRUE then calander months and ccamlr months will be generated}
}
\description{
Takes a dataframe and converts one of the columns to R date format and creates
the CCAMLR month (Dec=1, ... Nov=12)
}
