% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_tag_data}
\alias{process_tag_data}
\title{Process data for the multiple release function}
\usage{
process_tag_data(catch_data, release_data, recapture_data, location, species,
  select_catch = NULL, select_releases = NULL, select_recaptures = NULL,
  ...)
}
\arguments{
\item{catch_data}{CCAMLR C2 data extract}

\item{release_data}{CCAMLR tag release data extract}

\item{recapture_data}{CCAMLR tag release and recapture data extract}

\item{location}{either a ASD code or a research block (must be character format)}

\item{species}{either "TOA" or "TOP"}

\item{select_catch}{catch fields to select}

\item{select_releases}{tag release fields to select}

\item{select_recaptures}{tag recapture fields to select}

\item{...}{additional arguments}
}
\description{
Wrapper function over the three datasets, returns a list of three dataframes
}
