% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recapture_data_sim_RB}
\alias{recapture_data_sim_RB}
\title{Simulated tagged fish recapture data from hypothetical Research blocks for vignette examples}
\format{A dataframe with 12 rows and 7 columns
\describe{
\item{SEASON_RELEASE}{The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) the tagged fish was released}
\item{SEASON_RECAPTURE}{The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) the tagged fish was recaptured}
\item{CRUISE_ID_RECAPTURE}{An ID used by Observers that is recorded with tagged fish recaptures}
\item{SET_ID_RECAPTURE}{An ID used by Observers that is recorded with recaptured fish for each set-haul}
\item{SPECIES_CODE_RECAPTURE}{The three letter code used by CCAMLR to record the species of a tagged fish recapture. In the simulated data example this represent either Antarctic toothfish = "TOA" or Patagonian toothfish = "TOP"}
\item{RESEARCH_BLOCK_CODE_RECAPTURE}{A simulated hypothetical Research block code that the tagged fish recapture data fall within}
\item{RESEARCH_BLOCK_CODE_RELEASE}{A simulated hypothetical Research block code that the tagged fish release data fall within}
}}
\usage{
recapture_data_sim_RB
}
\description{
A dataset that aims to simulate the tagged fish recapture data necessary for toothfish local biomass estimates that is stored in CCAMLR's cdb database
}
\keyword{datasets}
