% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ref_area_seabed_area}
\alias{Ref_area_seabed_area}
\title{Reference Area Seabed Area}
\format{A dataframe with 6 rows and 2 columns
\describe{
\item{RefArea}{CCAMLR Reference area codes noting that RSR_open and HIMI are the only Reference areas currently used}
\item{Seabed_area}{Planimetric seabed area within the 600-1800 m estimated from a processed version of the GEBCO 2014 data - see inst/calculate_and_store_seabed_area_values.R}
}}
\usage{
Ref_area_seabed_area
}
\description{
A dataset that contains fishable seabed area within Reference Areas that were used in used in estimating local biomass at WG-FSA-17 (see Fig 1 in WG-FSA-17/42)
}
\keyword{datasets}
