% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{release_data_sim_RefArea}
\alias{release_data_sim_RefArea}
\title{Simulated tagged fish release data from Reference Areas for vignette examples}
\format{A dataframe with 200 rows and 7 columns
\describe{
\item{SEASON}{The simulated CCAMLR fishing season (i.e. December from the previous calendar year through to the end of November from the current calendar year) the tagged fish was released}
\item{CRUISE_ID}{An ID used by Observers that is recorded with tagged fish releases}
\item{SET_ID}{An ID used by Observers that is recorded for each released fish for each set-haul}
\item{SPECIES_CODE}{The three letter code used by CCAMLR to record a species. In the simulated data example this represent either Antarctic toothfish = "TOA" or Patagonian toothfish = "TOP"}
\item{REF_AREA_CODE}{A simulated Reference Area that the data fall within that represents the current Reference Areas for Antarctic and Patagonian toothfish species}
\item{ASD_CODE}{A simulated hypothetical CCAMLR Management area that represents an Area, Subarea or Division code that the data fall within}
\item{LENGTH_CM}{A simulated tagged and released toothfish length}
}}
\usage{
release_data_sim_RefArea
}
\description{
A dataset that aims to simulate the tagged fish release data necessary for toothfish local biomass estimates that is stored in CCAMLR's cdb database
}
\keyword{datasets}
