% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{release_weight_est}
\alias{release_weight_est}
\title{Estimate weight of released fish from length}
\usage{
release_weight_est(data)
}
\arguments{
\item{data}{Data object with all data extracts}
}
\description{
Extract catch data for cpue by seabed area biomass
}
