% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_ref.R
\name{removeBatchEffectRefs}
\alias{removeBatchEffectRefs}
\title{A method to remove batch effects estimated from a subset (references)
per batch only.
Source code is heavily based on limma::removeBatchEffects by
Gordon Smyth and Carolyn de Graaf}
\usage{
removeBatchEffectRefs(x, batch, references)
}
\arguments{
\item{x}{the data matrix with samples in columns and features in rows}

\item{batch}{the batch list as vector.}

\item{references}{a vector of integers, indicating whether the corresponding
sample is to be co-adjusted (0) or may be used as a reference (>0)}
}
\value{
the corrected data matrix
}
\description{
A method to remove batch effects estimated from a subset (references)
per batch only.
Source code is heavily based on limma::removeBatchEffects by
Gordon Smyth and Carolyn de Graaf
}
