% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_Data.R
\name{replace_missing}
\alias{replace_missing}
\title{Replaces missing values (NaN) by NA, this appears to be faster}
\usage{
replace_missing(data)
}
\arguments{
\item{data}{The data as dataframe}
}
\value{
The data with the replaced MVs
}
\description{
Replaces missing values (NaN) by NA, this appears to be faster
}
