% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{specify_params}
\alias{specify_params}
\title{Specify parameters for tag-based biomass estimate}
\usage{
specify_params(seasons, tag_mort, tag_shed, chronic_mort, report, nat_mort,
  move_rate)
}
\arguments{
\item{seasons}{vector of CCAMLR fishing seasons}

\item{tag_mort}{initial tag-induced mortality rates}

\item{tag_shed}{annual chronic (ongoing) tag shedding}

\item{chronic_mort}{vector of monthly chronic (ongoing) tag-induced mortality}

\item{report}{tag reporting rates}

\item{nat_mort}{annual natural mortality (instantaneous)}

\item{move_rate}{annual movement rate}
}
\description{
Specify parameters for tag-based biomass estimate
}
