% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Dataset.R
\name{strip_Covariable}
\alias{strip_Covariable}
\title{Strip column labelled Cov_1 from dataframe.}
\usage{
strip_Covariable(dataset)
}
\arguments{
\item{dataset}{Dataframe in the shape (samples, features) with additional
column Cov_1}
}
\value{
Dataset without column Cov_1.
}
\description{
Strip column labelled Cov_1 from dataframe.
}
