% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabed_area.R
\name{sum_area}
\alias{sum_area}
\title{Sum planimetric seabed area}
\usage{
sum_area(bathymetry)
}
\arguments{
\item{bathymetry}{is data in a raster file format upon whch the calculation
is based, this function is called in seabed_area}
}
\description{
Sum planimetric seabed area
}
