% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/multi_release.R,
%   R/single_release.R
\name{tag_bootstrap}
\alias{tag_bootstrap}
\alias{tag_bootstrap.srelease}
\alias{tag_bootstrap.mrelease}
\alias{tag_bootstrap.mrelease}
\alias{tag_bootstrap.srelease}
\title{Bootstrap method for single release tag-return studies}
\usage{
tag_bootstrap(x, nboot, boot_zeroes = TRUE)

\method{tag_bootstrap}{mrelease}(x, nboot, boot_zeroes = TRUE)

\method{tag_bootstrap}{srelease}(x, nboot, boot_zeroes = TRUE)
}
\arguments{
\item{x}{an object of class srelease}

\item{nboot}{number of bootstrap samples}

\item{boot_zeroes}{should bootstrap replicates with zero tag-recaptures be
used to estimate population size (default=TRUE)}
}
\description{
Bootstrap method for single release tag-return studies
}
\details{
Estimate confidence intervals using a non-parametric bootstrap. This method
incoporates uncertainty in tag-induced mortality, natural mortality and tag 
shedding as a series of Bernoulli trials. @seealso \code{\link{single_release}}
for more details.
}
