% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Dataset.R
\name{validate_input_generate_dataset}
\alias{validate_input_generate_dataset}
\title{Validate the user input to the function generate_dataset.
Raises an error if and only if the input is malformatted.}
\usage{
validate_input_generate_dataset(
  features,
  batches,
  samplesperbatch,
  mvstmt,
  classes,
  housekeeping,
  deterministic
)
}
\arguments{
\item{features}{Integer indicating the number of features
(e.g. genes/proteins) in the dataset.}

\item{batches}{Integer indicating the number of batches in the
dataset.}

\item{samplesperbatch}{Integer indicating the number of of samples
per batch.}

\item{mvstmt}{Float (in [0,1)) indicating the fraction of missing values
per batch.}

\item{classes}{Integer indicating the number of classes in the dataset.}

\item{housekeeping}{If NULL, no huosekeeping features will be simulatd.
Else, housepeeping indicates the fraction of of housekeeping features.}

\item{deterministic}{Whether to assigns the classes deterministically,
instead of random sampling}
}
\value{
None
}
\description{
Validate the user input to the function generate_dataset.
Raises an error if and only if the input is malformatted.
}
