% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_Data.R
\name{verify_references}
\alias{verify_references}
\title{Verify that the Reference column of the data contains only zeros and ones
(if it is present at all)}
\usage{
verify_references(batch)
}
\arguments{
\item{batch}{the dataframe for this batch (samples in rows, samples in columns)}
}
\value{
either TRUE (everything correct) or FALSE (something is not correct)
}
\description{
Verify that the Reference column of the data contains only zeros and ones
(if it is present at all)
}
