% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERT.R
\docType{package}
\name{BERT}
\alias{BERT}
\alias{BERT-package}
\title{BERT: Biomass Estimation for Research on Toothfish}
\description{
The BERT package provides methods for calculating precautionary estimates
of biomass using CPUE and tag-based methods to prevent stock depletion while
progressing towards an assessment.
}
\section{Multiple release, single release tag-return models}{

Implementation of Chapman estimators (Seber 1982)
along with bootstrapped confidence intervals
}

\section{CPUE seabed area}{

Implementation of the CPUE by seabed area analogy method (Agnew et al, 2009)
with bootstrapped confidence intervals
}

\section{Vignettes}{

To learn more about BERT, start with the vignette:
}

