% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/multi_release.R,
%   R/single_release.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.mrelease}
\alias{bootstrap.srelease}
\title{S3 generic bootstrap method}
\usage{
bootstrap(x, nboot, ...)

\method{bootstrap}{mrelease}(x, nboot, boot_zeroes = TRUE, ...)

\method{bootstrap}{srelease}(x, nboot, boot_zeroes = TRUE, ...)
}
\arguments{
\item{x}{an object of of the appropriate class}

\item{nboot}{number of bootstrap samples}

\item{...}{additional parameters}
}
\description{
Estimate confidence intervals using a non-parametric bootstrap.
}
