% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_catch_data_cpue_est}
\alias{extract_catch_data_cpue_est}
\title{Extract haul data}
\usage{
extract_catch_data_cpue_est(data, catch_seasons, measure, mean_fish_weight,
  target_species)
}
\arguments{
\item{data}{catch and tag recapture data
SEASON_RECAPTURE and Recapture_CCAMLR_Month}

\item{catch_seasons}{vector of seasons for which catch data exists}

\item{measure}{"numbers" if fish numbers are required or "weights" if fish weights are required}

\item{mean_fish_weight}{is the mean weight of a fish to estimate the fish release weights per haul from}

\item{target_species}{is the species that is targetted in the research block or reference area}
}
\description{
Extract catch data for cpue by seabed area biomass
}
