% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_catch_data_tag_est}
\alias{extract_catch_data_tag_est}
\title{Extract haul data}
\usage{
extract_catch_data_tag_est(data, rel_seasons, measure, mean_fish_weight)
}
\arguments{
\item{data}{catch and tag recapture data
SEASON_RECAPTURE and Recapture_CCAMLR_Month}

\item{rel_seasons}{vector of release seasons}

\item{measure}{"numbers" if fish numbers are required or "weights" if fish weights are required}

\item{mean_fish_weight}{is the mean weight of a fish to estimate the fish release weights per haul from}
}
\description{
Extract catch data for cpue by seabed area biomass
}
