% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_recaptures_season}
\alias{extract_recaptures_season}
\title{Extract and reformat release and recapture data}
\usage{
extract_recaptures_season(release_data, recapture_data,
  release_seasons = NULL, within_season_recaps = FALSE)
}
\arguments{
\item{release_data}{a single column data frame with the CCAMLR fishing season of individual fish release events}

\item{recapture_data}{a two column data frame with the CCAMLR fishing season of release in the first column with the matched CCAMLR fishing season of recapture in the second column for each individual tagged fish.}

\item{release_seasons}{vector of seasons of interest where the default = NULL uses release_data seasons included in release_data dataframe}

\item{within_season_recaps}{is TRUE to include within season recaptures or FALSE to exclude them with default=FALSE}
}
\description{
Extract and reformat recapture data for tag based biomass estimation
}
