% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert.R
\name{hierarchical_adjustment}
\alias{hierarchical_adjustment}
\title{Adjust data using the BERT algorithm.}
\usage{
hierarchical_adjustment(
  data,
  cores = 1,
  combatmode = 1,
  method = "ComBat",
  qualitycontrol = TRUE,
  verify = TRUE,
  mpi = FALSE,
  stopParLevel = FALSE
)
}
\arguments{
\item{data}{Matrix or dataframe in the format (samples, features).
Additional column names are "Batch", "Cov_X" (were X may be any number),
"Label" and "Sample".}

\item{cores}{The number of cores to use for parallel adjustment. Increasing
this number leads to faster adjustment, especially on Linux machines. The
default is 1.}

\item{combatmode}{Integer, encoding the parameters to use for ComBat.
1 (default)    par.prior = TRUE, mean.only = FALSE
2              par.prior = TRUE, mean.only = TRUE
3              par.prior = FALSE, mean.only = FALSE
4              par.prior = FALSE, mean.only = TRUE
Will be ignored, if method=="limma".}

\item{method}{Adjustment method to use. Should either be "ComBat" or "limma".}

\item{qualitycontrol}{Boolean indicating, whether ASWs should be computed before
and after batch effect adjustment. If TRUE, will compute ASW with respect to
the "Batch" and "Label" column (if existent).}

\item{verify}{Whether the input matrix/dataframe needs to be verified befire adjustment
(faster if FALSE)}

\item{mpi}{Whether to use MPI for inter-node parallelization.
Also allows None for testing purposes, which will perform no BE adjustment}

\item{stopParLevel}{Whether to stop parallelization at a certain hierarchy level.
FALSE by default. If integer, BERT will execute sequentially for all hierarchy levels= stopParLevel to avoid communication overhead.}
}
\value{
A matrix/dataframe mirroring the shape of the input. The data will
be batch-effect adjusted by BERT.
}
\description{
This function uses the hierarchical BERT algorithm to adjust data with
batch effects. It assumes that the data is in the format
(samples, features) and that missing values are indicated by NA.
An additional column labelled "Batch" should indicate the batch. Furthermore
all columns named "Cov_1", "Cov_2", ... will be considered as covariate
for adjustment. Columns labelled "Label" and "Sample" will be ignored,
all other columns are assumed to contain data.
}
