% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_bathy_data}
\alias{load_bathy_data}
\title{Load GEBCO bathymetry data}
\usage{
load_bathy_data(update = TRUE)
}
\arguments{
\item{update}{is TRUE if you want to load the most current bathymetry data from the CCAMLRGIS website or FALSE if you want to use the data that is stored in the package
(gebco 2014 30-arc second grid) will be downloaded and stored}
}
\description{
Load GEBCO bathymetry data
}
