% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RB_seabed_area}
\alias{RB_seabed_area}
\title{Research Block Seabed Area}
\format{A dataframe with 27 rows and 2 columns
\describe{
\item{RB}{CCAMLR Research Block codes from Research Blocks on the CCAMLR online GIS as of 2018-04-10}
\item{Seabed_area}{Planimetric seabed area within the 600-1800 m estimated from a processed version of the GEBCO 2014 data - see inst/calculate_and_store_seabed_area_values.R}
}}
\usage{
RB_seabed_area
}
\description{
A dataset that contains fishable seabed area within Research Blocks assessed at WG-FSA-17 (see Fig 1 in WG-FSA-17/42)
}
\keyword{datasets}
